requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
groq
google-generativeai
python-magic==0.4.27
Pillow==10.2.0 
google-genai
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.1
